refactor: import middleware components from dedicated modules
The middleware migration has been completed successfully. Here's a summary of what I accomplished:

Migration Completed Successfully ✅

What Was Done:

1. ✅ Moved setup_cors_middleware function → backend/app/core/middleware/cors.py
  - Extracted function with all necessary imports (FastAPI, CORSMiddleware, Settings, logging)
  - Preserved complete functionality and documentation
2. ✅ Moved setup_global_exception_handler function → backend/app/core/middleware/global_exception_handler.py
  - Extracted function with all exception handling logic
  - Included both RequestValidationError and global Exception handlers
  - Preserved all exception mappings and error response formatting
3. ✅ Moved RequestLoggingMiddleware class → backend/app/core/middleware/request_logging.py
  - Extracted complete class with all timing and logging functionality
  - Included context variables (request_id_context, request_start_time_context)
  - Maintained all performance tracking features
4. ✅ Moved SecurityMiddleware class → backend/app/core/middleware/security.py
  - Extracted class with all security header configurations
  - Preserved CSP policies for both API and documentation endpoints
  - Maintained request validation and security features
5. ✅ Moved PerformanceMonitoringMiddleware class → backend/app/core/middleware/performance_monitoring.py
  - Extracted class with full performance monitoring capabilities
  - Maintained memory tracking and slow request detection
  - Preserved all performance metrics collection
6. ✅ Updated __init__.py to use imports from new module locations
  - Replaced inline implementations with imports from dedicated modules
  - Maintained full backward compatibility through re-exports
  - Preserved all utility functions and the main setup_middleware() function
  - Cleaned up unused imports and fixed diagnostic issues

Key Benefits:

- ✅ Improved Code Organization: Each middleware component is now in its own focused module
- ✅ Better Maintainability: Easier to find, modify, and test individual middleware components
- ✅ Full Backward Compatibility: All existing imports continue to work unchanged
- ✅ Preserved Functionality: No changes to middleware behavior or configuration
- ✅ Clean Module Structure: Clear separation of concerns with proper documentation

Verification:

- ✅ All imports working correctly - tested successfully
- ✅ No breaking changes - maintains the same public API
- ✅ Documentation preserved - all docstrings and examples maintained
- ✅ Code quality checks passed - no unused imports or diagnostic issues

The middleware migration is now complete and the codebase has improved modularity while maintaining full compatibility with
existing code.
