+++ backend/app/core/middleware/__init__.py
@@ -113,792 +113,34 @@ Note:
 
 import logging
 import time
-import uuid
-from typing import Callable, Optional, Dict, Any
-from contextvars import ContextVar
-from starlette.types import ASGIApp, Receive, Scope, Send
+from typing import Optional, Dict
 
-from fastapi import FastAPI, Request, Response, status
-from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import JSONResponse
-from starlette.middleware.base import BaseHTTPMiddleware
-
+from fastapi import FastAPI, Request, Response
 from app.core.config import Settings
-from app.core.exceptions import (
-    ApplicationError,
-    ValidationError,
-    AuthenticationError,
-    AuthorizationError,
-    ConfigurationError,
-    BusinessLogicError,
-    InfrastructureError,
-    AIServiceException,
-    TransientAIError,
-    PermanentAIError,
-    get_http_status_for_exception
-)
-from fastapi.exceptions import RequestValidationError
-from app.schemas.common import ErrorResponse
+
+# Import middleware components from their dedicated modules
+from .cors import setup_cors_middleware
+from .global_exception_handler import setup_global_exception_handler
+from .request_logging import RequestLoggingMiddleware, request_id_context, request_start_time_context
+from .security import SecurityMiddleware
+from .performance_monitoring import PerformanceMonitoringMiddleware
 
 # Configure module logger
 logger = logging.getLogger(__name__)
 
-# Context variables for request tracking
-request_id_context: ContextVar[str] = ContextVar('request_id', default='')
-request_start_time_context: ContextVar[float] = ContextVar('request_start_time', default=0.0)
+# Context variables are now imported from request_logging module
+# but we re-export them here for backward compatibility
 
 
 # ============================================================================
-# CORS Middleware Configuration
+# Middleware Components - Imported from dedicated modules
 # ============================================================================
 
-def setup_cors_middleware(app: FastAPI, settings: Settings) -> None:
-    """
-    Configure Cross-Origin Resource Sharing (CORS) middleware for the application.
-    
-    Sets up CORS middleware with production-ready security settings that allow
-    controlled cross-origin access while preventing unauthorized requests. The
-    configuration supports development and production environments with appropriate
-    security controls.
-    
-    CORS Features:
-        * Configurable allowed origins from settings
-        * Support for credentials in cross-origin requests
-        * All HTTP methods allowed for API flexibility
-        * All headers allowed for maximum compatibility
-        * Preflight request handling for complex requests
-        * Security-conscious defaults with explicit configuration
-    
-    Args:
-        app (FastAPI): The FastAPI application instance to configure
-        settings (Settings): Application settings containing CORS configuration
-            including allowed_origins list
-    
-    Configuration:
-        The middleware uses the following settings:
-        * settings.allowed_origins: List of allowed origin URLs
-        * Credentials: Enabled to support authentication cookies/headers
-        * Methods: All HTTP methods (*) for maximum API flexibility
-        * Headers: All headers (*) for client compatibility
-    
-    Security Notes:
-        * Origins are explicitly configured, not using wildcard in production
-        * Credentials support requires specific origin configuration
-        * Preflight requests are properly handled for complex CORS scenarios
-        * Settings validation ensures only valid origins are configured
-    
-    Example:
-        >>> setup_cors_middleware(app, settings)
-        >>> # CORS middleware now configured with settings.allowed_origins
-    
-    Note:
-        CORS middleware should be added last in the middleware stack to ensure
-        it processes responses after all other middleware has completed. This
-        is handled automatically by the setup_middleware() function.
-    """
-    logger.info(f"Setting up CORS middleware with origins: {settings.allowed_origins}")
-    
-    # Note: Type ignore comments are needed due to FastAPI/Starlette type inference
-    # limitations with CORSMiddleware. The code functions correctly at runtime.
-    app.add_middleware(
-        CORSMiddleware,  # type: ignore[arg-type]
-        allow_origins=settings.allowed_origins,  # type: ignore[call-arg]
-        allow_credentials=True,  # type: ignore[call-arg]
-        allow_methods=["*"],  # type: ignore[call-arg]
-        allow_headers=["*"],  # type: ignore[call-arg]
-    )
-
-
-# ============================================================================
-# Global Exception Handler
-# ============================================================================
-
-def setup_global_exception_handler(app: FastAPI, settings: Settings) -> None:
-    """
-    Configure global exception handling for unhandled application errors.
-    
-    Provides a centralized error handling mechanism that catches all unhandled
-    exceptions across the application and returns consistent, secure error responses.
-    The handler ensures clients receive predictable error responses while protecting
-    internal implementation details and sensitive information.
-    
-    Exception Handling Features:
-        * Comprehensive logging of all unhandled exceptions with full context
-        * Standardized error response format using shared.models.ErrorResponse
-        * HTTP status code mapping based on exception types
-        * Security-conscious error messages that don't expose internal details
-        * Request context preservation for debugging and monitoring
-        * Integration with the custom exception hierarchy
-    
-    Args:
-        app (FastAPI): The FastAPI application instance to configure
-        settings (Settings): Application settings for error handling configuration
-    
-    Exception Processing:
-        The handler processes exceptions in the following order:
-        1. Log the full exception with context for debugging
-        2. Determine appropriate HTTP status code based on exception type
-        3. Generate secure, user-friendly error message
-        4. Create standardized ErrorResponse model
-        5. Return JSONResponse with appropriate status code
-    
-    HTTP Status Code Mapping:
-        * ApplicationError -> 400 Bad Request (validation, business logic errors)
-        * InfrastructureError -> 502 Bad Gateway (external service failures)
-        * TransientAIError -> 503 Service Unavailable (temporary AI issues)
-        * PermanentAIError -> 502 Bad Gateway (permanent AI issues)
-        * All other exceptions -> 500 Internal Server Error
-    
-    Security Features:
-        * Generic error messages prevent information disclosure
-        * Full exception details logged server-side only
-        * No stack traces or internal paths exposed to clients
-        * Request correlation IDs for secure debugging
-    
-    Example Response:
-        ```json
-        {
-            "success": false,
-            "error": "Internal server error",
-            "error_code": "INTERNAL_ERROR",
-            "timestamp": "2025-07-12T12:34:56.789012"
-        }
-        ```
-    
-    Note:
-        This handler is the last resort for exception handling and will catch
-        any exception not handled by more specific exception handlers. It ensures
-        the application never returns unhandled exceptions to clients.
-    """
-    
-    @app.exception_handler(RequestValidationError)
-    async def request_validation_exception_handler(request: Request, exc: RequestValidationError) -> JSONResponse:
-        """
-        Handle FastAPI request validation errors with consistent formatting.
-        
-        This handler catches Pydantic validation errors that occur during request
-        parsing and converts them to the standardized ErrorResponse format. This
-        ensures all validation errors (both automatic and custom) follow the same
-        response structure.
-        
-        Args:
-            request (Request): The FastAPI request object that failed validation
-            exc (RequestValidationError): The Pydantic validation error
-        
-        Returns:
-            JSONResponse: Standardized error response with validation details
-        """
-        # Get request ID for correlation (if available)
-        request_id = getattr(request.state, 'request_id', request_id_context.get('unknown'))
-        
-        # Extract validation error details
-        error_details = []
-        for error in exc.errors():
-            field_name = " -> ".join(str(loc) for loc in error["loc"])
-            error_details.append(f"{field_name}: {error['msg']}")
-        
-        error_message = "Invalid request data"
-        if error_details:
-            # Include first error for user feedback, log all details
-            error_message = f"Invalid request data: {error_details[0]}"
-        
-        # Log the validation error with context
-        logger.warning(
-            f"Request validation error in request {request_id}: {'; '.join(error_details)}",
-            extra={
-                'request_id': request_id,
-                'method': request.method,
-                'url': str(request.url),
-                'validation_errors': exc.errors()
-            }
-        )
-        
-        # Create standardized error response
-        error_response = ErrorResponse(
-            error=error_message,
-            error_code="VALIDATION_ERROR"
-        )
-        
-        return JSONResponse(
-            status_code=422,  # Unprocessable Entity (standard for validation errors)
-            content=error_response.dict()
-        )
-    
-    @app.exception_handler(Exception)
-    async def global_exception_handler(request: Request, exc: Exception) -> JSONResponse:
-        """
-        Global exception handler for unhandled application errors.
-        
-        Catches all unhandled exceptions, logs them appropriately, and returns
-        a standardized error response to protect internal implementation details
-        while providing useful information for debugging and monitoring.
-        
-        Args:
-            request (Request): The FastAPI request object that triggered the exception
-            exc (Exception): The unhandled exception that was raised during processing
-        
-        Returns:
-            JSONResponse: A standardized error response with appropriate HTTP status
-        """
-        # Get request ID for correlation (if available)
-        request_id = getattr(request.state, 'request_id', request_id_context.get('unknown'))
-        
-        # Log the full exception with context
-        logger.error(
-            f"Unhandled exception in request {request_id}: {str(exc)}",
-            extra={
-                'request_id': request_id,
-                'method': request.method,
-                'url': str(request.url),
-                'exception_type': type(exc).__name__,
-                'exception_module': type(exc).__module__
-            },
-            exc_info=True
-        )
-        
-        # Determine HTTP status code based on exception type
-        http_status = get_http_status_for_exception(exc)
-        
-        # Create standardized error response
-        error_response = ErrorResponse(
-            error="Internal server error",
-            error_code="INTERNAL_ERROR"
-        )
-        
-        # Customize error message based on exception type for better UX
-        if isinstance(exc, (ValidationError, BusinessLogicError)):
-            error_response.error = "Invalid request data"
-            error_response.error_code = "VALIDATION_ERROR"
-        elif isinstance(exc, AuthenticationError):
-            error_response.error = "Authentication failed"
-            error_response.error_code = "AUTHENTICATION_ERROR"
-        elif isinstance(exc, AuthorizationError):
-            error_response.error = "Access denied"
-            error_response.error_code = "AUTHORIZATION_ERROR"
-        elif isinstance(exc, ConfigurationError):
-            error_response.error = "Service configuration error"
-            error_response.error_code = "CONFIGURATION_ERROR"
-        elif isinstance(exc, TransientAIError):
-            error_response.error = "AI service temporarily unavailable"
-            error_response.error_code = "SERVICE_UNAVAILABLE"
-        elif isinstance(exc, PermanentAIError):
-            error_response.error = "AI service error"
-            error_response.error_code = "AI_SERVICE_ERROR"
-        elif isinstance(exc, InfrastructureError):
-            error_response.error = "External service error"
-            error_response.error_code = "INFRASTRUCTURE_ERROR"
-        
-        return JSONResponse(
-            status_code=http_status,
-            content=error_response.dict()
-        )
-
-
-# ============================================================================
-# Request Logging Middleware
-# ============================================================================
-
-class RequestLoggingMiddleware(BaseHTTPMiddleware):
-    """
-    HTTP request/response logging middleware with performance tracking.
-    
-    Provides comprehensive logging of HTTP requests and responses with timing
-    information, request correlation IDs, and configurable detail levels.
-    The middleware supports both development debugging and production monitoring
-    with appropriate log filtering and security considerations.
-    
-    Logging Features:
-        * Request/response timing with millisecond precision
-        * Unique request ID generation for correlation tracking
-        * HTTP method, URL, status code, and response size logging
-        * User agent and client IP address tracking (when available)
-        * Configurable log levels based on response status codes
-        * Sensitive data filtering for security compliance
-        * Integration with structured logging for monitoring systems
-    
-    Performance Tracking:
-        * Request start/end timing with high precision
-        * Response size calculation for bandwidth monitoring
-        * Slow request detection and warning logs
-        * Memory usage tracking (optional)
-        * Integration with performance monitoring systems
-    
-    Security Features:
-        * Request ID correlation for secure debugging
-        * Sensitive header filtering (Authorization, API keys)
-        * Query parameter sanitization for logs
-        * Client IP anonymization options
-        * Log level control to prevent information leakage
-    
-    Configuration:
-        The middleware behavior can be controlled through settings:
-        * request_logging_enabled: Enable/disable request logging
-        * log_level: Minimum log level for request logging
-        * slow_request_threshold: Threshold for slow request warnings
-        * sensitive_headers: List of headers to filter from logs
-    
-    Example Logs:
-        ```
-        INFO: Request started: GET /v1/text_processing/process [req_id: abc123]
-        INFO: Request completed: GET /v1/text_processing/process 200 1.2KB 150ms [req_id: abc123]
-        WARN: Slow request: POST /v1/text_processing/batch_process 200 5.6KB 2500ms [req_id: def456]
-        ```
-    
-    Note:
-        This middleware should be placed early in the middleware stack to capture
-        timing information for all subsequent middleware and request processing.
-    """
-    
-    def __init__(self, app: ASGIApp, settings: Settings):
-        """
-        Initialize request logging middleware with configuration.
-        
-        Args:
-            app (ASGIApp): The ASGI application to wrap
-            settings (Settings): Application settings for logging configuration
-        """
-        super().__init__(app)
-        self.settings = settings
-        self.slow_request_threshold = getattr(settings, 'slow_request_threshold', 1000)  # ms
-        self.sensitive_headers = {'authorization', 'x-api-key', 'cookie'}
-        
-    async def dispatch(self, request: Request, call_next: Callable[[Request], Any]) -> Response:
-        """
-        Process HTTP request with logging and timing.
-        
-        Args:
-            request (Request): The incoming HTTP request
-            call_next (Callable): The next middleware/handler in the chain
-            
-        Returns:
-            Response: The HTTP response from downstream handlers
-        """
-        # Generate unique request ID for correlation
-        request_id = str(uuid.uuid4())[:8]
-        request_id_context.set(request_id)
-        
-        # Store request ID in request state for access by other components
-        request.state.request_id = request_id
-        
-        # Record request start time
-        start_time = time.time()
-        request_start_time_context.set(start_time)
-        
-        # Log request start
-        logger.info(
-            f"Request started: {request.method} {request.url.path} [req_id: {request_id}]",
-            extra={
-                'request_id': request_id,
-                'method': request.method,
-                'path': request.url.path,
-                'query_params': dict(request.query_params),
-                'user_agent': request.headers.get('user-agent', 'unknown'),
-                'client_ip': request.client.host if request.client else 'unknown'
-            }
-        )
-        
-        try:
-            # Process request through remaining middleware and handlers
-            response = await call_next(request)
-            
-            # Calculate request duration
-            duration_ms = (time.time() - start_time) * 1000
-            
-            # Get response size if available
-            response_size = getattr(response, 'content-length', 'unknown')
-            if hasattr(response, 'body') and response.body:
-                response_size = f"{len(response.body)}B"
-            
-            # Determine log level based on status code
-            log_level = logging.INFO
-            if response.status_code >= 400:
-                log_level = logging.WARNING if response.status_code < 500 else logging.ERROR
-            elif duration_ms > self.slow_request_threshold:
-                log_level = logging.WARNING
-                
-            # Log request completion
-            logger.log(
-                log_level,
-                f"Request completed: {request.method} {request.url.path} "
-                f"{response.status_code} {response_size} {duration_ms:.1f}ms [req_id: {request_id}]",
-                extra={
-                    'request_id': request_id,
-                    'method': request.method,
-                    'path': request.url.path,
-                    'status_code': response.status_code,
-                    'duration_ms': duration_ms,
-                    'response_size': response_size
-                }
-            )
-            
-            return response
-            
-        except Exception as exc:
-            # Log request exception
-            duration_ms = (time.time() - start_time) * 1000
-            logger.error(
-                f"Request failed: {request.method} {request.url.path} "
-                f"{type(exc).__name__} {duration_ms:.1f}ms [req_id: {request_id}]",
-                extra={
-                    'request_id': request_id,
-                    'method': request.method,
-                    'path': request.url.path,
-                    'duration_ms': duration_ms,
-                    'exception_type': type(exc).__name__
-                },
-                exc_info=True
-            )
-            raise
-
-
-# ============================================================================
-# Security Middleware
-# ============================================================================
-
-class SecurityMiddleware(BaseHTTPMiddleware):
-    """
-    Security-focused middleware for HTTP headers and request validation.
-    
-    Provides essential security controls including security headers injection,
-    basic request validation, and protection against common web vulnerabilities.
-    The middleware implements defense-in-depth security principles suitable
-    for production API deployments.
-    
-    Security Features:
-        * Security headers injection (HSTS, X-Frame-Options, etc.)
-        * Content Security Policy (CSP) configuration
-        * X-Content-Type-Options and X-XSS-Protection headers
-        * Request size limits to prevent DoS attacks
-        * Basic input validation and sanitization
-        * Referrer policy configuration for privacy
-        * Feature policy headers for browser security
-    
-    Headers Configured:
-        * Strict-Transport-Security: HSTS for HTTPS enforcement
-        * X-Frame-Options: Clickjacking protection
-        * X-Content-Type-Options: MIME type sniffing prevention
-        * X-XSS-Protection: XSS attack mitigation
-        * Content-Security-Policy: Resource loading restrictions
-        * Referrer-Policy: Referrer information control
-        * Permissions-Policy: Browser feature access control
-    
-    Request Validation:
-        * Content-Length limits for request body size
-        * Header count and size validation
-        * URL length restrictions
-        * Basic malicious pattern detection
-        * Rate limiting hooks (basic implementation)
-    
-    Configuration:
-        Security behavior can be configured through settings:
-        * security_headers_enabled: Enable/disable security headers
-        * max_request_size: Maximum request body size in bytes
-        * max_headers_count: Maximum number of headers per request
-        * csp_policy: Custom Content Security Policy string
-    
-    Note:
-        This middleware provides basic security controls. For production
-        deployments, additional security measures like WAF, rate limiting,
-        and DDoS protection should be implemented at the infrastructure level.
-    """
-    
-    def __init__(self, app: ASGIApp, settings: Settings):
-        """
-        Initialize security middleware with configuration.
-        
-        Args:
-            app (ASGIApp): The ASGI application to wrap
-            settings (Settings): Application settings for security configuration
-        """
-        super().__init__(app)
-        self.settings = settings
-        self.max_request_size = getattr(settings, 'max_request_size', 10 * 1024 * 1024)  # 10MB
-        self.max_headers_count = getattr(settings, 'max_headers_count', 100)
-        
-        # Base security headers (applied to all endpoints)
-        self.base_security_headers = {
-            'Strict-Transport-Security': 'max-age=31536000; includeSubDomains',
-            'X-Frame-Options': 'DENY',
-            'X-Content-Type-Options': 'nosniff',
-            'X-XSS-Protection': '1; mode=block',
-            'Referrer-Policy': 'strict-origin-when-cross-origin',
-            'Permissions-Policy': 'geolocation=(), microphone=(), camera=()'
-        }
-        
-        # Strict CSP for API endpoints (production security)
-        self.api_csp = "default-src 'self'; script-src 'self'; style-src 'self'; img-src 'self' data:; font-src 'self'; connect-src 'self'; object-src 'none'; base-uri 'self'; frame-ancestors 'none';"
-        
-        # Relaxed CSP for documentation endpoints (Swagger UI compatibility)
-        self.docs_csp = "default-src 'self' https://fastapi.tiangolo.com; script-src 'self' 'unsafe-inline' 'unsafe-eval' https://cdn.jsdelivr.net https://unpkg.com; style-src 'self' 'unsafe-inline' https://cdn.jsdelivr.net https://unpkg.com https://fonts.googleapis.com; img-src 'self' https://fastapi.tiangolo.com data: https:; font-src 'self' https://cdn.jsdelivr.net https://unpkg.com https://fonts.gstatic.com; connect-src 'self'; object-src 'none'; base-uri 'self';"
-        
-        # Documentation endpoints that need relaxed CSP
-        self.docs_endpoints = {'/docs', '/redoc', '/openapi.json'}
-    
-    def _is_docs_endpoint(self, request: Request) -> bool:
-        """
-        Determine if the request is for a documentation endpoint.
-        
-        Checks if the request path matches any of the known documentation
-        endpoints that require relaxed CSP policies for Swagger UI functionality.
-        
-        Args:
-            request (Request): The incoming HTTP request
-            
-        Returns:
-            bool: True if the request is for a documentation endpoint
-        """
-        path = request.url.path
-        
-        # Check exact matches for docs endpoints
-        if path in self.docs_endpoints:
-            return True
-            
-        # Check for docs-related paths (e.g., /docs/oauth2-redirect)
-        if path.startswith('/docs') or path.startswith('/redoc'):
-            return True
-            
-        # Check for internal-related paths (e.g., /internal/docs)
-        if path.startswith('/internal'):
-            return True
-            
-        # Check for OpenAPI schema variants
-        if 'openapi' in path.lower():
-            return True
-            
-        return False
-        
-    async def dispatch(self, request: Request, call_next: Callable[[Request], Any]) -> Response:
-        """
-        Process HTTP request with security validations and header injection.
-        
-        Args:
-            request (Request): The incoming HTTP request
-            call_next (Callable): The next middleware/handler in the chain
-            
-        Returns:
-            Response: The HTTP response with security headers
-        """
-        # Validate request headers count
-        if len(request.headers) > self.max_headers_count:
-            logger.warning(f"Request with excessive headers: {len(request.headers)}")
-            return JSONResponse(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                content={"error": "Too many headers"}
-            )
-        
-        # Validate Content-Length if present
-        content_length = request.headers.get('content-length')
-        if content_length:
-            try:
-                size = int(content_length)
-                if size > self.max_request_size:
-                    logger.warning(f"Request too large: {size} bytes")
-                    return JSONResponse(
-                        status_code=status.HTTP_413_REQUEST_ENTITY_TOO_LARGE,
-                        content={"error": "Request too large"}
-                    )
-            except ValueError:
-                logger.warning("Invalid Content-Length header")
-                return JSONResponse(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    content={"error": "Invalid Content-Length"}
-                )
-        
-        # Process request through remaining middleware
-        response = await call_next(request)
-        
-        # Inject base security headers
-        for header, value in self.base_security_headers.items():
-            response.headers[header] = value
-        
-        # Determine and inject appropriate CSP based on endpoint type
-        if self._is_docs_endpoint(request):
-            response.headers['Content-Security-Policy'] = self.docs_csp
-            logger.debug(f"Applied docs CSP for {request.url.path}")
-        else:
-            response.headers['Content-Security-Policy'] = self.api_csp
-            logger.debug(f"Applied API CSP for {request.url.path}")
-            
-        return response
-
-
-# ============================================================================
-# Performance Monitoring Middleware
-# ============================================================================
-
-class PerformanceMonitoringMiddleware(BaseHTTPMiddleware):
-    """
-    Performance monitoring middleware for request timing and resource tracking.
-    
-    Provides comprehensive performance monitoring including request timing,
-    memory usage tracking, and performance metrics collection. The middleware
-    supports both real-time monitoring and historical performance analysis
-    with integration capabilities for external monitoring systems.
-    
-    Monitoring Features:
-        * High-precision request timing with nanosecond accuracy
-        * Memory usage tracking before and after requests
-        * CPU usage monitoring (when available)
-        * Slow request detection with configurable thresholds
-        * Request concurrency tracking
-        * Response size and bandwidth monitoring
-        * Performance metrics aggregation
-        * Integration hooks for external monitoring systems
-    
-    Metrics Collected:
-        * Request duration (total, processing, waiting)
-        * Memory usage (RSS, heap, available)
-        * CPU utilization during request processing
-        * Request/response sizes for bandwidth analysis
-        * Concurrent request counts
-        * Error rates and status code distributions
-        * Endpoint-specific performance patterns
-    
-    Performance Analysis:
-        * Real-time performance alerting for slow requests
-        * Statistical analysis of response times
-        * Memory leak detection through usage patterns
-        * Bottleneck identification in request processing
-        * Performance regression detection
-        * Resource usage optimization recommendations
-    
-    Configuration:
-        Monitoring behavior can be configured through settings:
-        * performance_monitoring_enabled: Enable/disable monitoring
-        * slow_request_threshold: Threshold for slow request alerts
-        * memory_monitoring_enabled: Enable memory usage tracking
-        * metrics_export_enabled: Enable metrics export to external systems
-    
-    Integration:
-        The middleware can export metrics to monitoring systems:
-        * Prometheus metrics export
-        * StatsD metrics publishing
-        * CloudWatch metrics integration
-        * Custom metrics webhooks
-    
-    Example Metrics:
-        ```
-        request_duration_seconds{method="POST", endpoint="/v1/text_processing/process"} 0.125
-        request_memory_usage_bytes{method="POST", endpoint="/v1/text_processing/process"} 1048576
-        slow_requests_total{method="POST", endpoint="/v1/text_processing/process"} 1
-        ```
-    
-    Note:
-        Performance monitoring adds minimal overhead but should be configured
-        appropriately for high-traffic production environments. Consider
-        sampling strategies for very high request volumes.
-    """
-    
-    def __init__(self, app: ASGIApp, settings: Settings):
-        """
-        Initialize performance monitoring middleware.
-        
-        Args:
-            app (ASGIApp): The ASGI application to wrap
-            settings (Settings): Application settings for monitoring configuration
-        """
-        super().__init__(app)
-        self.settings = settings
-        self.slow_request_threshold = getattr(settings, 'slow_request_threshold', 1000)  # ms
-        self.memory_monitoring_enabled = getattr(settings, 'memory_monitoring_enabled', True)
-        
-    async def dispatch(self, request: Request, call_next: Callable[[Request], Any]) -> Response:
-        """
-        Process HTTP request with performance monitoring.
-        
-        Args:
-            request (Request): The incoming HTTP request
-            call_next (Callable): The next middleware/handler in the chain
-            
-        Returns:
-            Response: The HTTP response with performance headers
-        """
-        import psutil
-        import os
-        
-        # Get request ID for correlation
-        request_id = getattr(request.state, 'request_id', 'unknown')
-        
-        # Record performance baseline
-        start_time = time.perf_counter()
-        start_memory = None
-        
-        if self.memory_monitoring_enabled:
-            try:
-                process = psutil.Process(os.getpid())
-                start_memory = process.memory_info().rss
-            except Exception:
-                # Memory monitoring is optional
-                pass
-        
-        try:
-            # Process request
-            response = await call_next(request)
-            
-            # Calculate performance metrics
-            duration = time.perf_counter() - start_time
-            duration_ms = duration * 1000
-            
-            # Memory usage calculation
-            memory_delta = None
-            if self.memory_monitoring_enabled and start_memory:
-                try:
-                    process = psutil.Process(os.getpid())
-                    end_memory = process.memory_info().rss
-                    memory_delta = end_memory - start_memory
-                except Exception:
-                    pass
-            
-            # Add performance headers
-            response.headers['X-Response-Time'] = f"{duration_ms:.2f}ms"
-            if memory_delta is not None:
-                response.headers['X-Memory-Delta'] = f"{memory_delta}B"
-            
-            # Log performance metrics
-            perf_extra = {
-                'request_id': request_id,
-                'method': request.method,
-                'path': request.url.path,
-                'duration_ms': duration_ms,
-                'status_code': response.status_code
-            }
-            
-            if memory_delta is not None:
-                perf_extra['memory_delta_bytes'] = memory_delta
-            
-            # Slow request detection
-            if duration_ms > self.slow_request_threshold:
-                logger.warning(
-                    f"Slow request detected: {request.method} {request.url.path} "
-                    f"{duration_ms:.1f}ms [req_id: {request_id}]",
-                    extra=perf_extra
-                )
-            else:
-                logger.debug(
-                    f"Performance: {request.method} {request.url.path} "
-                    f"{duration_ms:.1f}ms [req_id: {request_id}]",
-                    extra=perf_extra
-                )
-            
-            return response
-            
-        except Exception as exc:
-            # Log performance for failed requests
-            duration = time.perf_counter() - start_time
-            duration_ms = duration * 1000
-            
-            logger.error(
-                f"Performance (failed): {request.method} {request.url.path} "
-                f"{duration_ms:.1f}ms {type(exc).__name__} [req_id: {request_id}]",
-                extra={
-                    'request_id': request_id,
-                    'method': request.method,
-                    'path': request.url.path,
-                    'duration_ms': duration_ms,
-                    'exception_type': type(exc).__name__
-                }
-            )
-            raise
+# setup_cors_middleware function is imported from cors.py
+# setup_global_exception_handler function is imported from global_exception_handler.py
+# RequestLoggingMiddleware class is imported from request_logging.py
+# SecurityMiddleware class is imported from security.py
+# PerformanceMonitoringMiddleware class is imported from performance_monitoring.py
 
 
 # ============================================================================
@@ -1039,7 +281,7 @@ def get_request_id(request: Request) -> str:
     return getattr(request.state, 'request_id', request_id_context.get('unknown'))
 
 
-def get_request_duration(request: Request) -> Optional[float]:
+def get_request_duration(_request: Request) -> Optional[float]:
     """
     Get the current request duration in milliseconds.
     
@@ -1047,7 +289,7 @@ def get_request_duration(request: Request) -> Optional[float]:
     for performance monitoring and timeout detection within request handlers.
     
     Args:
-        request (Request): The FastAPI request object
+        _request (Request): The FastAPI request object (unused, timing obtained from context)
         
     Returns:
         Optional[float]: Request duration in milliseconds, or None if timing
